🩹 fix(settings): share invitation member type
