+++ apps/webapp/src/app/[locale]/(app)/settings/employees/employees-page-client.tsx
@@ -36,26 +36,16 @@ import {
 	TableRow,
 } from "@/components/ui/table";
 import { useCompilerSafeReactTable } from "@/components/use-compiler-safe-react-table";
-import type * as authSchema from "@/db/auth-schema";
-import type { employee } from "@/db/schema";
+import type {
+	InvitationWithInviter,
+	MemberWithUserAndEmployee,
+} from "@/components/organization/people-management-types";
 import { useEmployeeClockStatuses } from "@/lib/query";
 import { useEmployees } from "@/lib/query/use-employees";
 import type { SettingsAccessTier } from "@/lib/settings-access";
 import { Link } from "@/navigation";
 import { columns } from "./columns";
 import type { EmployeeDirectoryRow } from "./employee-action-types";
-
-export interface MemberWithUserAndEmployee {
-	member: typeof authSchema.member.$inferSelect;
-	user: typeof authSchema.user.$inferSelect;
-	employee: typeof employee.$inferSelect | null;
-	teamMemberships?: Array<{ teamId: string }>;
-}
-
-export type InvitationWithInviter = typeof authSchema.invitation.$inferSelect & {
-	user: typeof authSchema.user.$inferSelect;
-	targetTeam?: { id: string; name: string } | null;
-};
 
 export interface EmployeesPagePeopleProps {
 	organizationName: string;
